queryLatest: replicate.upto -> createLogStream
Replace use of replicate.upto as `ssb-db2` does not appear to respond
properly to requests using that muxrpc. Instead, use `createLogStream`
to achieve the same thing albeit at the expense of some performance.

Also implemented is a small heuristic to ensure that the javascript & go
implementations (which have slightly different responses when answering
a createLogStream call) all work.